20230205 2213 Update
Started organising the code in Game Chooser and added the class Game_Chooser that will run the main code.
Basically finished most of the functionalities. I could put more time and refine it further, but I just want to move on to the next lesson now. Perhaps will look into adding more features and refining in the future.